Ajout de la page des jeux, et donc différente de la page d'accueil
