+++ src/app/pages/games/games-routing.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { GamesPage } from './games.page';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: GamesPage
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class GamesPageRoutingModule {}
