+++ src/app/pages/games/games.module.ts
@@ -1,0 +1,20 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+
+import { IonicModule } from '@ionic/angular';
+
+import { GamesPageRoutingModule } from './games-routing.module';
+
+import { GamesPage } from './games.page';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    FormsModule,
+    IonicModule,
+    GamesPageRoutingModule
+  ],
+  declarations: [GamesPage]
+})
+export class GamesPageModule {}
