+++ src/app/pages/games/games.page.scss
@@ -1,0 +1,100 @@
+#desc{
+    padding:5px;
+    text-align: center;
+    display: none;
+  }
+  #page{
+    background-color:transparent; 
+   border:1px solid rgb(230,230,230);
+   text-align:left;
+   width:200px;
+   height: 100%;
+   min-height: 230px;
+   margin: auto;
+  }
+  
+  h1{
+   font-size:30px;
+   font-family:Arial, sans serif;
+  }
+  
+  img.pendu{
+   float:center;
+   width:170px;
+   margin: auto;
+  }
+  .out{
+   //float:right;
+    font-size:16px;
+    padding: 10px;
+    text-align: center;
+  }
+  
+  #nbre_essai{
+  font-weight:bold;
+  color:#e64e02;
+  }
+  
+  h2 {
+   font-size:16px;
+   font-family:Arial, sans-serif;
+  }
+  
+  #mot {
+   margin:0px auto 4px auto;
+   clear:both; 
+  }
+  
+  #mot td {
+   padding:6px;
+   background-color:#e64e02;
+   border:1px solid rgb(230,230,230);
+   width:30px;
+   height:30px;
+   font-weight:bold;
+   font-size:1.5em;
+   text-align:center;
+   color:white;
+  }
+  
+  #mot span{
+   visibility:hidden;
+  }
+  
+  #clavier{
+   border:1px solid rgb(200,200,200);
+   padding:10px;
+   background-color:rgb(225,225,225);
+   width: 100%;
+  }
+  
+  #clavier tr{
+    width:100%;
+  }
+  
+  #clavier td {
+   background-color:white;
+   border:1px solid rgb(200,200,200);
+   width:14.2%;
+   height:30px;
+   text-align:center;
+   font-weight:bold;
+   cursor:pointer;
+   border-radius:2px;
+   -webkit-border-radius:2px;
+   -moz-border-radius:2px;
+   -ms-border-radius:2px;
+   -o-border-radius:2px;
+  }
+  
+  a.lien{
+   color:blue;
+   text-decoration:none;
+   font-size:20px;
+  }
+  
+  a.lien:hover{
+   color:red;
+   font-weight:bold;
+   text-decoration:underline;
+  }
